pr #1349. Updated documentation and fixed broken tests.

+++ src/psyclone/tests/test_files/dynamo0p3/15.22.4_int_inc_a_minus_X_builtin.f90
@@ -44,7 +44,7 @@ program single_invoke
   implicit none
 
   type(integer_field_type) :: f1
-  real(i_def)              :: a
+  integer(i_def)           :: a
 
   a = 3
